Resolve portal/discovery/mols_test.go in favor of the compressed suite:
main's #320 generalized the MOLS internals tests to the dynamic-order
implementation, but this PR removes those internals pins entirely per
review; the kept selection-contract tests and benchmarks are identical
on both sides and pass against the new GF(N^2) implementation.

+++ portal/telemetry/trace.go
@@ -7,11 +7,11 @@ import "time"
 type SelectionTrace struct {
 	Timestamp time.Time
 
-	// ClientHash is hashToGF64(LocalAddress) — a single byte derived from the
-	// client identity. Used only for sampled debug-log correlation. Not a
+	// ClientHash is the FNV-1a hash of LocalAddress folded into the grid order
+	// at selection time. Used only for sampled debug-log correlation. Not a
 	// Prometheus label (would unbounded cardinality) and not a public field on
 	// any external API.
-	ClientHash uint8
+	ClientHash uint32
 
 	// Mode is "priority" or "multihop", matching the calling method.
 	Mode string
